+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## v0.0.2
+
+### FEATURES
+
+- Implemented `discover releases` command ([#1])
+
+### IMPROVEMENTS
+
+- Detect Cosmos chains based on go dependency ([#2])
+- Show number of forks in discover projects table view ([#4])
+- Allow sorting projects by number of forks or stars ([#5])
+
+### BUG FIXES
+
+[#1]: https://github.com/andynog/cosmocope/issues/1
+[#2]: https://github.com/andynog/cosmocope/issues/2
+[#4]: https://github.com/andynog/cosmocope/issues/4
+[#5]: https://github.com/andynog/cosmocope/issues/5
+
 ## v0.0.1
 *March 7, 2021*